pyproject.toml
[project]
name = "dutato-curriculum-creator"
version = "0.1.0"
description = "Create structured curricula from diverse source materials, designed for AI agent operation"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.25.0",
    "tiktoken>=0.9.0",
    "supabase>=2.15.0",
    "rich>=13.9.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "python-dotenv>=1.1.0",
    "trafilatura>=2.0.0",
]
